+++ common/dao-api/src/main/java/org/thingsboard/server/dao/mes/vo/TanSensorDeviceRunVo.java
@@ -24,7 +24,7 @@ public class TanSensorDeviceRunVo extends BaseRowModel {
     private String deviceName;
     @ExcelProperty(value = "运行时间",index = 4)
     @ApiModelProperty("运行时间")
-    private Long runSeund;
+    private BigDecimal runSeund;
     @ExcelProperty(value = "最高温度",index = 5)
     @ApiModelProperty("最高温度")
     private BigDecimal maxTemp;
